+++ vrc-get-vpm/src/io/tokio.rs
@@ -184,8 +184,16 @@ impl<T: TokioIoTraitImpl + Sync> IoTrait for T {
         fs::remove_file(self.resolve(path)?).await
     }
 
+    async fn remove_dir(&self, path: &Path) -> io::Result<()> {
+        fs::remove_dir(self.resolve(path)?).await
+    }
+
     async fn remove_dir_all(&self, path: &Path) -> io::Result<()> {
         fs::remove_dir_all(self.resolve(path)?).await
+    }
+
+    async fn rename(&self, from: &Path, to: &Path) -> io::Result<()> {
+        fs::rename(self.resolve(from)?, self.resolve(to)?).await
     }
 
     async fn metadata(&self, path: &Path) -> io::Result<Metadata> {