+++ vrc-get-vpm/src/repository/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod local;
 mod remote;
 
+pub use local::LocalCachedRepository;
 pub use remote::RemotePackages;
 pub use remote::RemoteRepository;